added functional testing for page caching.  need more comprehensive tests to find that caching bug...
git-svn-id: http://svn.techno-weenie.net/projects/mephisto/trunk@859 567b1171-46fb-0310-a4c9-b4bef9110e78

+++ app/cachers/template_sweeper.rb
@@ -1,7 +1,14 @@
 class TemplateSweeper < ActionController::Caching::Sweeper
-  observe Template
+  observe Template, LayoutTemplate
+
+  # only sweep updates, not creations
+  # tagged 'lame hack'
+  def before_save(record)
+    @new = record.new_record?
+  end
 
   def after_save(record)
+    return if @new
     controller.class.benchmark "Expired all referenced pages" do
       CachedPage.find(:all).each { |p| controller.class.expire_page(p.url) }
       CachedPage.delete_all
